+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Reduce bioformats source memory usage ([#1038](../../pull/1038))
 - Better pick the largest image from bioformats ([#1039](../../pull/1039), [#1040](../../pull/1040))
 - Cache histogram thresholds ([#1042](../../pull/1042))
+- Add `_repr_png_` for Jupyter ([#1058](../../pull/1058))
 
 ## 1.19.3