+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Render item lists faster ([#1036](../../pull/1036))
 - Reduce bioformats source memory usage ([#1038](../../pull/1038))
 - Better pick the largest image from bioformats ([#1039](../../pull/1039), [#1040](../../pull/1040))
+- Speed up some images handled with bioformats ([#1063](../../pull/1063))
 - Cache histogram thresholds ([#1042](../../pull/1042))
 - Add `_repr_png_` for Jupyter ([#1058](../../pull/1058))
 - Ignore bogus tifffile resolutions ([#1062](../../pull/1062))